refactor: change structure files
